rename base.json and base.js files to tokens.json and tokens.js to improve consistency
