+++ tokens/components/hint/tokens.json
