+++ tokens/components/label/tokens.json
