+++ tokens/global/outline/tokens.json
